Initial Commit
The third of the first 3 mods on schedule for Rhynia (mod for Nodecore). Intended to replace the old "nc_eggave", though with a different effect. Like it's other 2 partner plant mods, it is a shell of it's former self, as Rhynia is intended to take care of much of the basics that all old nc_flora plants redundantly had coded for themselves in all mods. This plant in particular may change more so than the others as the old "nc_liana" mod will likely be rewritten as a second genus within this mod in keeping with the simultaneous occurrence of both genera together (though this plant will no longer have it's thorny effects (barring someone makes an actually working status effect mod for Nodecore)).

This plant was recreated when an example plant was needed to demonstrate the difference between the growth rate of plants with and without the growth_opt trait in Rhynia. And thus should remain that way in marginally better keeping with a xerophytic plant.